pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meetingscribe"
version = "1.0.0"
description = "Privacy-first meeting transcription and summarization — local Whisper, speaker diarization, AI summaries"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Zachary Gray", email = "zacharytgray@gmail.com" },
]
keywords = [
    "meeting",
    "transcription",
    "summarization",
    "whisper",
    "diarization",
    "privacy",
    "speech-to-text",
    "ai",
    "claude",
    "openrouter",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "faster-whisper>=1.0.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.24,<2",
    "anthropic>=0.25.0",
    "pyannote.audio>=3.1.0",
    "torch>=2.0.0",
    "pystray>=0.19.4",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
macos = ["pyobjc-framework-Cocoa>=10.0"]

[project.urls]
Homepage = "https://github.com/zacharytgray/MeetingScribe"
Repository = "https://github.com/zacharytgray/MeetingScribe"
"Bug Tracker" = "https://github.com/zacharytgray/MeetingScribe/issues"

[project.scripts]
meetingscribe = "meetingscribe.cli_entry:main"
meetingscribe-tray = "meetingscribe.cli_entry:tray_main"
meetingscribe-app = "meetingscribe.cli_entry:app_main"

[tool.setuptools.packages.find]
where = ["."]
include = ["meetingscribe*"]
